+++ CHANGELOG.md
@@ -11,3 +11,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - v1beta metrics: get total listens
 - v1beta metrics: get total users
 - v1beta metrics: get user listens leaderboard
+- v1beta users: get user by username, user ID, discord ID, and URI
+- v1beta users: get user profile
+- v1beta users: get user history stats (total listens, total artists)
+- v1beta users: get recently played
+- v1beta users: get currently playing